LayoutTests/fast/dom/DeviceOrientation/add-listener-from-callback-expected.txt

Tests that adding a new event listener from a callback works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS deviceOrientationEvent.alpha is mockAlpha
PASS deviceOrientationEvent.beta is mockBeta
PASS deviceOrientationEvent.gamma is mockGamma
PASS deviceOrientationEvent.alpha is mockAlpha
PASS deviceOrientationEvent.beta is mockBeta
PASS deviceOrientationEvent.gamma is mockGamma
PASS deviceOrientationEvent.alpha is mockAlpha
PASS deviceOrientationEvent.beta is mockBeta
PASS deviceOrientationEvent.gamma is mockGamma
PASS successfullyParsed is true

TEST COMPLETE
